[parsing] Rename detail_scene_graph ==> detail_sdf_geometry
Marginally relevant to: #16229

Whatever sense the orginal name made, it no longer expresses what is in
the module, nor how it differs from other modules here.
